fix: 9 bugs from the 2026-07-11 full-codebase review
- drum-grid visual-yank crash: range clipboard now u64/64-wide, matching
  step_count's 32->64 grow
- group FX-chain stale-retarget: deleting a group now drops (not just
  leaves dangling) any undo entry or open nudge batch naming that slot,
  so the next addGroup reusing it can't get silently overwritten
- multiband compressor crossover load-order: new MultibandComp.setXovers
  sets both points from a trusted-valid saved pair instead of cross-
  clamping lo against hi's stale pre-load default
- ':'/'/' from synth/sampler/spectrum (no arm of their own) now flush
  pending param/FX nudge batches via the shared mode_changed handler
- orphaned sidecar WAVs: save() now prunes stale <stem>_samples/ files
  left behind when a track/pad's index shifts
- swapTracks now sends all_notes_off like every other track-mutating
  Session method
- spectrum accumulator reset now also fires on a source-category change
  (track<->group), not just a numeric-index change, since both share
  track_spectrum's one accumulator
- MIDI CC dirty flag no longer fires when the command queue was full and
  the event was actually dropped
- BandComp.gainFor now reuses Compressor's envelope/ratio math instead of
  a hand-copied duplicate

+++ src/tui/editors/drum.zig
@@ -447,7 +447,7 @@ fn yankSelection(app: *App) void {
         var s: u8 = r.lo;
         while (s <= r.hi) : (s += 1) {
             if (!dm.stepActive(@intCast(pad), s)) continue;
-            const bit = @as(u32, 1) << @intCast(s - r.lo);
+            const bit = @as(u64, 1) << @intCast(s - r.lo);
             clip.active[pad] |= bit;
             clip.vel[pad][s - r.lo] = dm.stepVel(@intCast(pad), s);
         }
@@ -488,7 +488,7 @@ fn pasteSelection(app: *App) void {
         const target = base +| i;
         if (target >= dm.step_count) break;
         for (0..DrumMachine.max_pads) |pad| {
-            const bit = @as(u32, 1) << @intCast(i);
+            const bit = @as(u64, 1) << @intCast(i);
             const active = clip.active[pad] & bit != 0;
             setStep(dm, @intCast(pad), target, active, clip.vel[pad][i]);
         }
